Updating Font

+++ HTML - Peter Choi/js/controllers/app-logout.js
@@ -1,13 +1,20 @@
 angular
   .module("myApp")
   .controller("LogoutCtrl", LogoutCtrl)
-
-  function LogoutCtrl() {
+  LogoutCtrl.$inject = ['$window']
+  function LogoutCtrl($window) {
 
     var self = this;
 
   self.loggingOut = function() {
-    console.log("THE FUNCTION!!!")
-    sessionStorage.clear();
+    var r = confirm("Are you sure you want to log out?");
+    if(r==true) {
+      console.log("confirmed logout")
+      sessionStorage.clear();
+        $window.location.href = '#/';
+    } else {
+      console.log("canceled logout")
+    }
+
 }
 }
